Move the WorkExperience.jsx to a separate folder 'workExperience'.
