+++ src/components/form/ui/Form.jsx
@@ -4,7 +4,7 @@ import PersonalInformation from "../components/PersonalInformation";
 import SocialMedia from "../components/SocialMedia";
 import Summary from "../components/Summary";
 import Education from "../components/Education";
-import WorkExperience from "../components/WorkExperience";
+import WorkExperience from "../components/workExperience/ui/WorkExperience";
 import Projects from "../components/projects/ui/Projects";
 import Skills from "../components/skills/ui/Skills";
 import Languages from "../components/languages/ui/Languages";
